requirements.txt
pandas>=1.1.3
matplotlib
numpy
scipy
speedtest-cli
isort
python-papi
setuptools==59.6.0
scikit-learn
PyQt6
requests